refactor: simplify collision, input, and sprite systems
The collision system is now passed the player entity as an argument. Now, when any computer entity
enters the world, we add a collider between it and the player sprite. The input system should not
need to care that the entity has a velocity component. Sprites are now set to collide with the world
bounds when they enter the world, rather than in the collision system.